добавляет компонент пустого списка офферов
